requirements.txt
# === BACKEND FRAMEWORK ===
fastapi==0.115.*
uvicorn==0.34.*
pydantic==2.*
python-dotenv==1.*

# === AI / LLM ===
langchain==0.3.*
langchain-google-genai>=2.0.0
langchain-community==0.3.*

# === VECTOR DB ===
chromadb==0.6.*

# === ĐỌC TÀI LIỆU ===
pypdf==5.*
python-docx==1.*

# === DATABASE ===
sqlalchemy==2.*

# === AUTH ===
firebase-admin==6.*

# === TIỆN ÍCH ===
httpx==0.28.*
python-multipart==0.0.*
